시크릿 베이스 페이지 작업중

+++ src/pages/secret-bases.vue
@@ -21,10 +21,10 @@
         <q-btn
           color="primary"
           label="재료 등록"
-          @click="showSecretBaseCreateDialog = true"/>
+          @click="secretBasePageStore.showCreateSecretBaseDialog"/>
       </q-card-section>
-      <q-card-section class="q-gutter-md flex">
 
+      <q-card-section class="q-gutter-md flex">
         <SecretBaseCard v-for="secretBase in filteredSecretBases" :key="secretBase.name"
                         :secret-base="secretBase"
                         @onUpdateButtonClick="onUpdateButtonClick"
@@ -34,46 +34,36 @@
 
     </BaseCard>
 
-    <CreateSecretBaseCard v-model="pageStore.showSecretBaseCreateDialog"/>
-    <UpdateSecretBaseCard v-model="pageStore.showSecretBaseUpdateDialog" :secret-base="updateSecretBase"/>
+    <CreateSecretBaseCard v-model="secretBasePageStore.createSecretBaseDialog"/>
+<!--    <UpdateSecretBaseCard v-model="secretBasePageStore.updateSecretBaseDialog"/>-->
   </q-page>
 </template>
 
+<script setup lang="ts">
+import {computed, ref} from 'vue';
+import {useSecretBaseStore} from 'stores/secret-base';
+import SecretBaseCard from 'components/apps/secret-base/SecretBaseCard.vue';
+import CreateSecretBaseCard from 'components/apps/secret-base/CreateSecretBaseCard.vue';
+import BaseCard from 'components/BaseCard.vue';
+import {useSecretBasePageStore} from 'stores/pages/secret-bases';
+import UpdateSecretBaseCard from 'components/apps/secret-base/UpdateSecretBaseCard.vue';
 
-
-
-<script setup>
-
-import {computed, ref} from "vue";
-import {useSecretBaseStore} from "stores/secret-base";
-import SecretBaseCard from "components/apps/secret-base/SecretBaseCard.vue";
-import CreateSecretBaseCard from "components/apps/secret-base/CreateSecretBaseCard.vue";
-import BaseCard from "components/BaseCard.vue";
-import UpdateSecretBaseCard from "components/apps/secret-base/UpdateSecretBaseCard.vue";
-import {useSecretBasePageStore} from "stores/pages/secret-base";
+const secretBaseStore = useSecretBaseStore();
+const secretBasePageStore = useSecretBasePageStore();
 
 const searchText = ref('');
-const secretBaseStore = useSecretBaseStore();
-
 const filteredSecretBases = computed(() => {
   return secretBaseStore.secretBases.filter(secretBase => {
     return secretBase.name.includes(searchText.value);
   });
 });
 
-const pageStore = useSecretBasePageStore();
-// const showSecretBaseCreateDialog = ref(false);
-// const showSecretBaseUpdateDialog = ref(false);
-// const updateSecretBase = ref(null);
 
-const showSecretBaseUpdate = () => showSecretBaseUpdateDialog.value = true;
-const onUpdateButtonClick = secretBase => {
-  pageStore.updateSecretBase.value = secretBase;
-  showSecretBaseUpdate();
-}
+
 const onDeleteButtonClick = secretBase => {
-  secretBaseStore.delete(secretBase);
-  alert('삭제되었습니다.');
+  if (confirm('정말 삭제 하시겠어요?')) {
+    secretBaseStore.delete(secretBase);
+  }
 }
 
 </script>
